Organized Components for Article Controls to match the directory structure of the other components.
