+++ packages/frontend-web/src/app/components/ArticleControls/index.tsx
@@ -1,0 +1,3 @@
+import {ArticleControlIcon} from './ArticleControls';
+
+export default ArticleControlIcon;
